requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
openai==1.86.0
faiss-cpu==1.11.0
langchain==0.3.25
langchain-openai==0.3.23
pydantic==2.33.2
pytest==8.4.0
python-dotenv==1.1.0
python-multipart==0.0.20